Symbol files alignment with english, and further fixes to luxembourgish (#19256)
Link to issue number:

Not applicable
Summary of the issue:

Updates the symbol files to align it with English latest changes
Description of user facing changes:

Adds hyphen with the corresponding fixes.
Description of developer facing changes:

Not applicable
Description of development approach:

Added the hyphen from the english file to be transferred to the speech synthesizer.