Update NVDA license information in README (#19251)
#19003 follow up 

### Summary of the issue:

This pull request makes a minor update to the licensing information in
the project documentation. The change clarifies that NVDA is available
under a modified GNU General Public License version 2 or any later
version.

Updated the license statement in `readme.md` to specify "version 2 or
later" rather than just "version 2" for the GNU General Public License.