fix crowdin sync to beta by running pyright locally (#19252)
Summary of the issue:

Fetching translations from Crowdin is failing because pre-commit is failing to install on the github actions image.
This is because pyright is failing to install.
We rely on pre-commit to ensure translated files don't introduce regressions.
https://github.com/nvaccess/nvda/actions/runs/19559152241
Description of user facing changes:

none
Description of developer facing changes:

run pyright with uv directly when using pre-commit. This might result in slightly different output to the official hook.

This has the following benefits:

    we don't need to manually sync versions
    it matches closer to what we run in GitHub CI (ci/scripts/tests/typeCheck.ps1), as pre-commit CI doesn't support pyright anyway

Description of development approach:

See above