Fix disabling then enabling touch support (#19280)
### Summary of the issue:
On 64-bit versions of NVDA, disabling touch support then reenabling it
causes an error. This subsequently renders the settings dialog
inoperable.

### Description of user facing changes:
Disabling and re-enabling touch support works correctly.

### Description of developer facing changes:
NOne

### Description of development approach:
Cast the window class attom to `LPCWSTR` before passing to
`UnregisterClass`, as ctypes does not know that the integer atom is a
valid value in this case.

### Testing strategy:
Built a self-signed launcher, installed it, then disabled and re-enabled
touch support via the settings dialog.

### Known issues with pull request:
None