Fix error when trying to read documents with malformed URL in links (#19289)
Fixes #19125
Summary of the issue:

In Chromium browsers, when the cursor reaches a link with a malformed URL, NVDA cannot read the document.
Description of user facing changes:

Fixed regression: It Will be posible to read documents in Chromium browsers, even if the contain links with a malformed URL.
Description of developer facing changes:

None
Description of development approach:

In the isSamePageURL function of urlUtils, if a ValueError exception is produced when parsing an URL, make the function to return False.