Fork repository CI does not fail when exiting early. (#19307)
Follow-up to #19254
Summary of the issue:

In the early exit step, the Action of Fork still ends in failure because the gh run command does not know the repository information.

This is similar to cli/cli#10525
Description of user facing changes:
Description of developer facing changes:
Description of development approach:

Add the GH_REPO environment variable for the early exit action step. Make the gh command find the repository.
Testing strategy:

Manual test: set GH_REPO on the command line and run gh run watch to get Action status successfully.