Move settings to Privacy and Security category (#19296)
Resolves #16272

### Summary of the issue:

NVDA has several settings that have a direct impact on users' privacy
and security, but they are scattered throughout its settings dialog.

### Description of user facing changes:

Some of these settings have now been moved to the Privacy and Security
category.

### Description of developer facing changes:

A few items have been removed from the public API.

### Description of development approach:

- Move the code for the logging level setting from
`GeneralSettingsPanel` to `PrivacyAndSecuritySettingsPanel`, and
slightly refactored it.
- Moved the code for the allow usage stats setting from
`GeneralSettingsPanel` to `PrivacyAndSecurityPanel`, and refactored
slightly.
- Searched the User Guide for references to logs/logging and
stats/statistics, and updated as appropriate.

### Testing strategy:

Tried changing log levels and ensured that the changes were reflected in
the log output and in the GUI.

Disabled the source check in `updateCheck`, and tried checking for an
update with usage stats gathering on and off, and checked that the
setting was reflected in the URL.

### Known issues with pull request:

Users may find it difficult to adjust to the change in location