Add support for Word footnote and endnote reference navigation (#19310)
Closes #19300 
Summary of the issue:

There is no way to navigate between footnote and endnote references in Word. IN UIA, they are reported as links but quick navigation to links does not work.
Description of user facing changes:

    Added references to the elements list for Word with and without UIA
    Added an unassigned quicknav command for references.

I chose the word reference here because footnotes/endnotes might be misleading, i.e. you're not navigating between the notes or towards them, you are navigating to the references that point at the foot/endnotes.
Description of developer facing changes:

N/a
Description of development approach:

Use the footnote/endnote annotation for UIA and the footnotes/endnotes collection in the object model.