When update is not available, do not remove corresponding GUI controls but disable them (#19332)
Closes #19320
Summary of the issue:

Update check is not available when running from source or when running in secure mode. In these cases, related controls are just not created in the GUI.
Description of user facing changes:

Update related controls are now created in the GUI but are disabled when there is no update check.
Description of developer facing changes:

N/A
Description of development approach:

    Created "Automatically check for updates to NVDA", "Notify for pending update on startup", "Allow NV Access to gather NVDA usage statistics" and the "Update mirror" group but disable them.
    When necessary, force also their value to denote that there is no update check / stats collection.
    No change log item for this minor modification