Add warnings to AI image descriptions (#19327)
Related to #19053
Related to #19298

### Summary of the issue:

The on-device AI image descriptions introduced into NVDA tend to
experience halucinations, especially when used on material other than
photographs.

### Description of user facing changes:

* Descriptions retrieved with this feature are now prefixed with "Could
be"
* The settings and temporary enable dialogs now warn the user that the
feature is experimental and not to use it in high-stakes situations
* The User Guide includes a slightly longer disclaimer about the feature

### Description of developer facing changes:

None

### Description of development approach:

Used a template string when outputting the AI slop so that a
hedge-phrase can be used.
Wrote up a warning message for the user guide. Shortened it slightly and
inserted into the settings panel and temporary enable dialog.

I have not updated `changes.md`, as that would just create merge
conflicts. I will do it in #19319.

### Testing strategy:

Ran NVDA. Checked that the settings panel speaks the warning and that
the warning is visible. Checked that the warning is shown in the temp
enable dialog.

### Known issues with pull request:

Doesn't address the underlying issue.
The language used for this feature also seems confusing and
inconsistent, but that is out of scope for this issue.