Don't play spelling error reporting sounds when typing if speech mode is on-demand or off (#19348)
Fixes #19323

### Summary of the issue:
Sounds reporting errors while typing were emitted even when speech mode
was off or on-demand.

### Description of user facing changes:
Spelling error sounds while typing are no longer emitted when speech
mode are on-demand or off.
### Description of developer facing changes:
N/A
### Description of development approach:
Check the config to play the error wave file only only if we ar not in
off or on-demand mode, i.e. only if we ar in "talk" or "beeps" mode.

### Testing strategy:
Manual test in the 4 speech modes.

### Known issues with pull request:
None