Set the MathCAT output code automatically by NVDA language (#19368)
Closes #19233.
Summary of the issue:

nemeth is unsuitable as a default MathCAT output code for NVDA's global user base (see #19233 for details).
Description of how this pull request fixes the issue:

Adds a new "automatic" Braille code config option and switched to it by default. This option:

    Checks if there's a reasonable default code to use for the current NVDA language (es -> CMU, sv -> Swedish, vi -> Vietnamese);
    Otherwise, falls back to ASCIIMath.

Testing strategy:

Changed NVDA language to Spanish and verified that CMU became the effective code when the Braille code option was set to auto.