NVDA+F1 now opens the log viewer even when logging is disabled (#19347)
Closes #19318

### Summary of the issue:
The log cannot be opened when log level is on Disabled.
### Description of user facing changes:
The log can be opened when log level is on Disabled.
### Description of developer facing changes:
N/A
### Description of development approach:
When log is disabled, just open the log as if we had clicked the
corresponding NVDA menu item.
### Testing strategy:
Manual test with:
* Changing the log level in the settings dialog
* starting with `-l 100` flag

### Known issues with pull request:
No test performed with `--no-logging` flag due to #19333.