Fixed an error that occurred when enabling visual highlighter for elements with a height less than 0. #19384
Fixes: #19383
Summary of the issue:

The highlighter gets an error getting a rectangle of objects with heights less than 0. This is because we are not handling the ValueError exception thrown by locationHelper.RectLTRB.
Description of user facing changes:

The highlighter does not report an error when it encounters an object with a height less than 0.
Description of developer facing changes:
Description of development approach:

Handle ValueError exception when NVDAHighlighter.updateContextRect gets the context rectangle