Fix Firefox browse mode not announcing aria-label when label element contains only hidden content (#19409)
Summary of the issue:

When a form control (such as a checkbox or radio button) has both:

    An aria-label attribute providing the accessible name
    A <label for="..."> element whose content is entirely aria-hidden

NVDA's browse mode in Firefox fails to announce the accessible name. The name is correctly announced in focus mode.

Example HTML that reproduces the issue:

<input type="checkbox" id="testinput" aria-label="TestArea">
<label for="testinput"><span aria-hidden="true">Test</span></label>

In Firefox browse mode, navigating to this checkbox announces only "check box not checked" with no name, even though the accessible name is "TestArea" from aria-label.

The root cause is in getLabelInfo() which determines if a control's label is "visible" by checking IA2_RELATION_LABELLED_BY targets. It only checked if the label element had STATE_SYSTEM_INVISIBLE, but a label element can be visible while containing only aria-hidden content, meaning it contributes no accessible text to the virtual buffer. This caused NVDA to incorrectly assume the user would read the label content in browse mode, so it didn't force the aria-label name to be announced.

It could be argued that exposing the labelled by association even if the name is set using aria-label is a bug in Firefox, since the label element is no longer contributing to the accessible name of the element and that it should be fixed there. I'm willing to open a Firefox bug for this if preferred.
Description of user facing changes:

Form controls with aria-label will now correctly have their accessible name announced in browse mode, even when an associated <label for> element exists but contains only hidden content.